+++ sekoia-io-xdr/constants.py
@@ -3,10 +3,12 @@ INTEGRATION_NAME = "SEKOIAOperationCenter"
 BASE_URL = "https://app.sekoia.io"
 
 ALERTS_V1_BASE_URL = f"{BASE_URL}/api/v1/sic/alerts"
-CUSTOM_PRIORITIES_V1_BASE_URL = f"{BASE_URL}/api/v1/sic/custom_priorities"
-CUSTOM_STATUSES_V1_BASE_URL = f"{BASE_URL}/api/v1/sic/custom_statuses"
 
 ASSETS_V1_BASE_URL = f"{BASE_URL}/api/v1/asset-management/assets"
 ASSETS_V2_BASE_URL = f"{BASE_URL}/api/v2/asset-management/assets"
 
 CASES_V1_BASE_URL = f"{BASE_URL}/api/v1/sic/cases"
+
+CUSTOM_PRIORITIES_V1_BASE_URL = f"{BASE_URL}/api/v1/sic/custom_priorities"
+CUSTOM_STATUSES_V1_BASE_URL = f"{BASE_URL}/api/v1/sic/custom_statuses"
+CUSTOM_VERDICTS_V1_BASE_URL = f"{BASE_URL}/api/v1/sic/custom_verdicts"